// src/domain/calidad/models/calidadHuertasData.model.ts
export interface CalidadHuertasData {
  IdRegistroEmpacado: number;
  IdLote: number;
  Fecha: Date;
  NoRegistro: string;
  Huerta: string;
  Municipio: string;
  Cajas: number;
  Peso: number;
  VelocidadCepillado: string;
  TipoCorte: string;
  EstadoFruta: string;
  Destino: string;
  Activo:boolean;
  Creador: string;
}